+++ pkgs/test/CHANGELOG.md
@@ -9,6 +9,8 @@
   in `GithubReporter`.
 * Only include VM debugger in output when using interactive debugging, suppress
   it when only using coverage.
+* Use a redirect html file for browser tests to avoid leaking websocket details
+  through process starting arguments.
 
 ## 1.31.2